seven day forecast

+++ app/src/main/java/com/example/weatherforecast/data/models/network/Coord.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Coord(
     @SerializedName("lat")
-    val lat: Double,
+    val lat: Float,
     @SerializedName("lon")
-    val lon: Double
+    val lon: Float
 )
